// ProcMgt/ClientApp/src/app/models/quotationrequestdetails.ts
export class Quotationrequestdetails {

  quotationRequestDetailID: string;
  quotationRequestHeaderID: string;
  MakeID: string;
  ModelID: string;
  ItemID: string;
  MeasurementUnitID: string;
  unitPrice: number;
  Quantity: number;
  grossAmount: number;
  netAmount: number;
  discountAmount: number;
  quotationValidDate: string;
  attachment: string; 
  
  itemDescription: string;
  measurementUnitName: string;
  makeName: string;
  modelName: string;

  pounitPrice: number;
  qty: number;
  totalAmount: number;

 
}